llvmpipe: bump dirty tracker to 64-bits.
We need a lot more dirty states for task/mesh

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/23066>

+++ src/gallium/drivers/llvmpipe/lp_context.h
@@ -114,7 +114,7 @@ struct llvmpipe_context {
 
    bool queries_disabled;
 
-   unsigned dirty; /**< Mask of LP_NEW_x flags */
+   uint64_t dirty; /**< Mask of LP_NEW_x flags */
    unsigned cs_dirty; /**< Mask of LP_CSNEW_x flags */
    /** Mapped vertex buffers */
    ubyte *mapped_vbuffer[PIPE_MAX_ATTRIBS];
